refac: use trait
