+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
     "account_traits_derive",
     "bytemuck",
     "cli",
+    "clients/rust/common",
     "clients/rust/restaking_client",
     "clients/rust/vault_client",
     "core",
@@ -55,6 +56,7 @@ futures = "0.3.31"
 jito-bytemuck = { path = "bytemuck", version = "=0.0.3" }
 jito-account-traits-derive = { path = "account_traits_derive", version = "=0.0.3" }
 jito-jsm-core = { path = "core", version = "=0.0.3" }
+jito-restaking-client-common = { path = "clients/rust/common", version = "=0.0.3" }
 jito-restaking-client = { path = "clients/rust/restaking_client", version = "=0.0.3" }
 jito-restaking-core = { path = "restaking_core", version = "=0.0.3" }
 jito-restaking-program = { path = "restaking_program", version = "=0.0.3" }
