+++ cli/Cargo.toml
@@ -21,6 +21,7 @@ env_logger = { workspace = true }
 jito-bytemuck = { workspace = true }
 jito-jsm-core = { workspace = true }
 jito-restaking-client = { workspace = true }
+jito-restaking-client-common = { workspace = true }
 jito-restaking-core = { workspace = true }
 jito-vault-client = { workspace = true }
 jito-vault-core = { workspace = true }
