+++ clients/rust/common/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod log;
