+++ clients/rust/restaking_client/src/log/ncn.rs
@@ -1,4 +1,5 @@
-use super::{account_header, field, section_header, PrettyDisplay};
+use jito_restaking_client_common::log::{account_header, field, section_header, PrettyDisplay};
+
 use crate::accounts::Ncn;
 
 impl PrettyDisplay for Ncn {
@@ -34,8 +35,9 @@ impl PrettyDisplay for Ncn {
 #[cfg(test)]
 mod tests {
     use anchor_lang::prelude::Pubkey;
+    use jito_restaking_client_common::log::PrettyDisplay;
 
-    use crate::{accounts::Ncn, log::PrettyDisplay};
+    use crate::accounts::Ncn;
 
     #[test]
     fn test_ncn_pretty_display_structure() {
