+++ clients/rust/restaking_client/src/log/ncn_vault_ticket.rs
@@ -1,4 +1,5 @@
-use super::{account_header, field, section_header, PrettyDisplay};
+use jito_restaking_client_common::log::{account_header, field, section_header, PrettyDisplay};
+
 use crate::accounts::NcnVaultTicket;
 
 impl PrettyDisplay for NcnVaultTicket {
@@ -24,8 +25,9 @@ impl PrettyDisplay for NcnVaultTicket {
 #[cfg(test)]
 mod tests {
     use anchor_lang::prelude::Pubkey;
+    use jito_restaking_client_common::log::PrettyDisplay;
 
-    use crate::{accounts::NcnVaultTicket, log::PrettyDisplay, types::SlotToggle};
+    use crate::{accounts::NcnVaultTicket, types::SlotToggle};
 
     #[test]
     fn test_ncn_vault_ticket_pretty_display_structure() {
