+++ clients/rust/vault_client/Cargo.toml
@@ -19,6 +19,7 @@ idl-build = ["anchor-lang/idl-build"]
 anchor-lang = { workspace = true }
 borsh = { workspace = true }
 bytemuck = { workspace = true }
+jito-restaking-client-common = { workspace = true }
 num-derive = { workspace = true }
 num-traits = { workspace = true }
 serde = { workspace = true }
